Upgrade Elixir to v1.5.1
Use the Elixir image directly, let it decide which Erlang version to
depend on.

We used to force the Erlang version to 19.2, because at the time
(2016.07), 19.3 just shipped and I didn't trust a .0 release. Now that
Erlang 20.0.4 is out, and we've been stress-testing it at RabbitMQ for
months, I know that I can trust it.

re #1
[#150763329]