Dial down open file limit per nginx worker
This is the limit per worker, and since the workers are set to the
number of CPUs, we currently run 8 workers, which means that the total
file limit is ~82k. If nginx has more than ~82k files open, we
definitely want to know about it.

nginx has currently 200 files open